requirements.txt
PySide6>=6.6
faster-whisper>=1.0
sounddevice>=0.4
numpy>=1.24
requests>=2.31
edge-tts>=6.1
pyttsx3>=2.90
psutil>=5.9
pywin32>=306
vosk>=0.3.45
piper-tts>=1.2
silero-vad>=5.1
pycaw>=20230407
mss>=9.0
Pillow>=10.0
duckduckgo-search>=6.0
chromadb>=0.5
nvidia-ml-py>=12.0
# ML layer (no new install needed — Ollama handles the embedding model)
# Run once: ollama pull nomic-embed-text
